pyproject.toml
[project]
name = "test-llm-with-crosswords"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jupyter>=1.1.1",
    "openai==1.97.1",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.7",
]
